Separated class file, updated examples
